gnu: Remove python-potr.
The last release of the package dates from 2018.
According to its self-description: "This software is experimental and
potentially insecure. Do not rely on it"
It depends on the cryptography library python-pycrypto, which has had
its last release in 2013 and "is unmaintained, obsolete, and contains
security vulnerabilities" according to its homepage.

* gnu/packages/python-crypto.scm (python-potr): Delete variable.