gnu: sugar: Augment GI_TYPELIB_PATH, don't replace it.
This is important so that the typelib files of libraries that may have been
propagated by Sugar activities are visible when launching the activities
via sugar-toolkit-gtk3's "sugar-activity3" tool.

Activities propagate their needed gobject introspection libraries; when
installed to the system profile these libraries are available to the Sugar
launchers.

* gnu/packages/sugar.scm (sugar)[arguments]: Do not replace GI_TYPELIB_PATH
but augment it.