gnu: ungoogled-chromium: Don't add absolute references to libGL and libvulkan.
Keeping track of the uses of these libraries is a lot of work and ultimately
not worth the trouble -- especially since Chromium supports multiple libGL's
with different ABIs.  Instead we take the easy route and rely on a search path.

* gnu/packages/chromium.scm (ungoogled-chromium)[arguments]: Remove patching
of libGL.so and libvulkan.so.  Instead wrap with LD_LIBRARY_PATH.