gnu: gcc-11: Adapt patching for hurd to not use gexp's.
As this causes issues with gcc-cross-boot0, since it doesn't use gexp's and
that means the builder script ends up broken.

I'm not testing this properly, I've only confirmed that with this change, you
can generate a derivation for hello for i586-gnu where that derivation and all
inputs don't have broken builder scripts. Plus the output for hello on
x86_64-linux is unchanged.

* gnu/packages/gcc.scm (gcc-11)[arguments]: Rewrite to not use gexps.
[native-inputs]: Add a patch for the hurd if that's the target.