gnu: alex4: Remove non-free package.
Its readme.txt goes out of its way to licence only ‘source code’ under
the GPL2+, a reading supposedly confirmed by the author:

https://libregamewiki.org/Talk:Alex_the_Allegator_4

However, we don't need to rely on the authenticity of that claim: the
game data is clearly not licenced, and is required to run the binary.
Without it, the game segfaults.

* gnu/packages/games.scm (alex4): Remove variable.

Fixes <https://issues.guix.gnu.org/65881>.