gnu: librsvg: Update to 2.56.4.
* gnu/packages/gnome.scm (librsvg): Update to 2.56.4.
[source]: Remove snippet.
[cargo-inputs]: Add rust-anyhow-1, rust-clap-complete-4,
rust-encoding-rs-0.8, rust-system-deps-6, rust-thiserror-1. Replace
rust-cairo-rs-0.15 with 0.17, rust-clap-2 with 4, rust-cssparser-0.28
with 0.29, rust-data-url-0.1 with 0.2, rust-gdk-gdkixbuf-0.15 with 0.17,
rust-gio-0.15 with 0.17, rust-glib-0.15 with 0.17, rust-markup5ever-0.10
with 0.11, rust-nalgebra-0.29 with 0.32, rust-pango-0.15 with 0.17,
rust-pangocairo-0.15 with 0.17, rust-rctree-0.4 with 0.5,
rust-selectors-0.23 with 0.24, rust-xml5ever-0.16 with 0.17.  Remove
rust-encoding-0.2.
[cargo-development-inputs]: Add rust-cast-0.3, rust-pango-0.17,
rust-pangocairo-0.17.  Replace rust-cairo-rs-0.15 with 0.17,
rust-criterion-0.3 with 0.4, rust-glib-0.15 with 0.17, rust-lopdf-0.26
with 0.29, rust-yeslogic-fontconfig-sys-3 with 4.  Remove
rust-test-generator-0.3.
[arguments]: Adjust custom phase 'loosen-test-boundaries to changes in
the test suite.