gnu: Remove unneeded libxml2 from native-inputs. (gtk-doc)
The gtk-doc wrapped binaries use XML_CATALOG_FILES to perform XML catalog
lookup and come with docbook-xsl within the variable which fixes the
docbook-xsl stylesheet lookup related failures with these packages.
This makes it unnecessary to include libxml2 as the cause of error was usually
with the gtk-doc binaries.

* gnu/packages/gnome.scm (gupnp-igd, dee, zeitgeist, libnma)[native-inputs]:
Remove libxml2.
* gnu/packages/rpm.scm (libmodulemd)[native-inputs]: Ditto.
* gnu/packages/security-token.scm (libu2f-host)[native-inputs]: Ditto.

Signed-off-by: Maxim Cournoyer <maxim.cournoyer@gmail.com>