gnu: commencement: Make ‘glibc-final’ immune to ‘%current-target-system’.
Previously, the value memoized for ‘glibc-final’ was influenced by that
of ‘%current-target-system’ since it is used as a default argument to
‘libc-for-target’.  Thus, due to the order in which modules were loaded
versus the time at which ‘%current-target-system’ was set, it was
possible to end up with ‘glibc-final’ set to the “wrong” libc, for
instance when doing:

  guix build -f guix.scm --target=i586-pc-gnu

from another project such as the Shepherd.

This is similar to the problem fixed in
625a847d2a554168bec6e294888d9d1f8ab03d08.

* gnu/packages/commencement.scm (glibc-final-with-bootstrap-bash)
(glibc-final): Explicitly pass (%current-system) as the first argument
to ‘libc-for-target’ so that the value of ‘%current-target-system’ at
that time has no influence.

Change-Id: I2f05898b8f9d1f99d4c75108441c7ad38728225f