gnu: fftw-cmake: New Variables.
This is a clone of the fftw function, but uses cmake to build in order
for the FFTW3LibraryDepends.cmake file to be built, neccesary for
packages with cmake builds that depend on fftw3.

See: https://bugzilla.redhat.com/show_bug.cgi?id=1729652#c5

This variable inherits from the existing fftw3 in order to not break any
downstream packages that make use of the target optimizations given by
the gnu build.

* gnu/packages/algebra.scm (fftw-cmake): New variable.

Change-Id: Idb06f86031123de3b2621953a385f4f57d9b5cc5
Signed-off-by: Sharlatan Hellseher <sharlatanus@gmail.com>