gnu: krita: Update to 5.2.1.
Due to a missing patch in the upstream libraqm, we cannot depend on the
existing libraqm library, but must instead use the one embedded in the
package. It compiles, but needs to be told to use the -fPIC flag.

Dependency fftw also does not provide the required CMake file
FFTW3Librarydepends.cmake, and only does so when compile via cmake and
not gnu. Due to the many packages that depend on fftw, a seperate
fftw-cmake was used.

* gnu/packages/kde.scm (krita): Update to 5.2.1.

Change-Id: I956451d05d8499580b909e843fd8e4418e396642
Signed-off-by: Sharlatan Hellseher <sharlatanus@gmail.com>